create RegressionTestHelper.cs
